Update install-ca-certs
While `install-ca-certs` allows for either `pem` or `crt` files, if no `pem` file is found it will attempt to install `*.pem` which more than likely doesn't exist. This causes a failure when installing a new `crt` file (i.,e., `corporate.crt`) without a corresponding `pem`.

This change simply checks whether `$cert_file` exists and if not, it continues to the next loop iteration.